+++ tests/examples/examples_test.go
@@ -164,7 +164,7 @@ func TestAccHelmLocal(t *testing.T) {
 				t *testing.T, stackInfo integration.RuntimeValidationStackInfo,
 			) {
 				assert.NotNil(t, stackInfo.Deployment)
-				assert.Equal(t, 15, len(stackInfo.Deployment.Resources))
+				assert.Equal(t, 16, len(stackInfo.Deployment.Resources))
 			},
 		})